+++ test/webhook.js
@@ -375,10 +375,14 @@ describe('Webhooks', function() {
 	});
 
 	it('member.added', async function() {
-		await Promise.all([
+		const [events] = await Promise.all([
 			promises.waitForEvents('member.added'),
 			chan.addMembers([jaapID]),
 		]);
+		const event = events[0];
+		expect(event).to.not.be.null;
+		expect(event.type).to.eq('member.added');
+		expect(event.user.id).to.eq(jaapID);
 	});
 
 	it('member.updated', async function() {
@@ -545,6 +549,17 @@ describe('Webhooks', function() {
 		expect(event.user.id).to.eq(jaapID);
 		expect(event.target_user).to.be.an('object');
 		expect(event.target_user.id).to.eq(thierryID);
+	});
+
+	it('channel.truncate webhook fires', async function() {
+		const [events] = await Promise.all([
+			promises.waitForEvents('channel.truncated'),
+			chan.truncate(),
+		]);
+		const event = events[0];
+		expect(event).to.not.be.null;
+		expect(event.type).to.eq('channel.truncated');
+		expect(event.channel).to.be.an('object');
 	});
 
 	it('channel.deleted', async function() {